pyproject.toml
[project]
name = "FashionSeach"
version = "0.1.0"
description = "Fashion Search Engine"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pgvector>=0.4.1",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "boto3>=1.40.71",
    "typer>=0.20.0",
    "torch>=2.9.0",
    "torchvision>=0.24.0",
    "open-clip-torch>=3.2.0",
    "tqdm>=4.67.1",
    "sentence-transformers>=5.1.2",
]
